fix: migrate frontend to typescript
