+++ package.json
@@ -51,6 +51,17 @@
     ]
   },
   "devDependencies": {
-    "nodemon": "^2.0.7"
+    "@types/bcryptjs": "^2.4.2",
+    "@types/cors": "^2.8.12",
+    "@types/express": "^4.17.13",
+    "@types/jsonwebtoken": "^8.5.5",
+    "@types/node": "^16.10.2",
+    "@types/react": "^17.0.25",
+    "@types/react-dom": "^17.0.9",
+    "@types/react-helmet": "^6.1.2",
+    "@types/react-router-dom": "^5.3.0",
+    "@types/validator": "^13.6.3",
+    "nodemon": "^2.0.7",
+    "typescript": "^4.4.3"
   }
 }
