--- src/components/Navbar/Navbar.jsx
@@ -1,9 +1,0 @@
-import React from 'react';
-import './Navbar.css';
-import {nav as Nav} from './nav';
-const navbar=(props)=>{
-    return(
-        <Nav pathname={props.pathname}/>  
-    );
-};
-export default navbar;
