+++ src/components/Navbar/Navbar.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+import "./Navbar.css";
+import { nav as Nav } from "./nav";
+const navbar = (props: any) => {
+  return <Nav pathname={props.pathname} />;
+};
+export default navbar;
