+++ src/react-app-env.d.ts
@@ -1,0 +1,2 @@
+declare module "*.css";
+declare module "*.scss";
